Add RISC OS filetypes to the files in the 500 directory.
The filetypes had been lost, it seems, when the files were uploaded to
Git. This change introduces the filetypes to all the files that I can
recognise by sight. There are some that are clearly data (or code),
but I don't know if that was their original filetype.

The changes have only been applid to the 500 directory, which appears
to be 32bit (not RISC OS 5-specific), and therefore the most useful.